// src/store/vendor/vendor.enum.ts
enum VendorActionTypeEnum {
  GET_VENDOR_SUCCESS = "GET_VENDOR_SUCCESS",
  GET_VENDOR_FAILED = "GET_VENDOR_FAILED",
  GET_VENDOR_PENDING = "GET_VENDOR_PENDING",

  GET_VENDOR_STOCK_HISTORY_SUCCESS = "GET_VENDOR_STOCK_HISTORY_SUCCESS",
  GET_VENDOR_STOCK_HISTORY_FAILED = "GET_VENDOR_STOCK_HISTORY_FAILED",
  GET_VENDOR_STOCK_HISTORY_PENDING = "GET_VENDOR_STOCK_HISTORY_PENDING",

  GET_VENDOR_ORDERS_PENDING = "GET_VENDOR_ORDERS_PENDING",
  GET_VENDOR_ORDERS_FAILED = "GET_VENDOR_ORDERS_FAILED",
  GET_VENDOR_ORDERS_SUCCESS = "GET_VENDOR_ORDERS_SUCCESS",

  GET_VENDOR_PRODUCT_PENDING = "GET_VENDOR_PRODUCT_PENDING",
  GET_VENDOR_PRODUCT_SUCCESS = "GET_VENDOR_PRODUCT_SUCCESS",
  GET_VENDOR_PRODUCT_FAILED = "GET_VENDOR_PRODUCT_FAILED",

  UPDATE_VENDOR_PRODUCT_PENDING = "UPDATE_VENDOR_PRODUCT_PENDING",
  UPDATE_VENDOR_PRODUCT_SUCCESS = "UPDATE_VENDOR_PRODUCT_SUCCESS",
  UPDATE_VENDOR_PRODUCT_FAILED = "UPDATE_VENDOR_PRODUCT_FAILED",

  ADD_VENDOR_PRODUCT_PENDING = "ADD_VENDOR_PRODUCT_PENDING",
  ADD_VENDOR_PRODUCT_SUCCESS = "ADD_VENDOR_PRODUCT_SUCCESS",
  ADD_VENDOR_PRODUCT_FAILED = "ADD_VENDOR_PRODUCT_FAILED",

  ADD_VENDOR_PRODUCT_TIER_PENDING = "ADD_VENDOR_PRODUCT_TIER_PENDING",
  ADD_VENDOR_PRODUCT_TIER_SUCCESS = "ADD_VENDOR_PRODUCT_TIER_SUCCESS",
  ADD_VENDOR_PRODUCT_TIER_FAILED = "ADD_VENDOR_PRODUCT_TIER_FAILED",

  DELETE_VENDOR_PRODUCT_PENDING = "DELETE_VENDOR_PRODUCT_PENDING",
  DELETE_VENDOR_PRODUCT_SUCCESS = "DELETE_VENDOR_PRODUCT_SUCCESS",
  DELETE_VENDOR_PRODUCT_FAILED = "DELETE_VENDOR_PRODUCT_FAILED",

  ADD_PRODUCT_TIER_PRICING_PENDING = "ADD_PRODUCT_TIER_PRICING_PENDING",
  ADD_PRODUCT_TIER_PRICING_SUCCESS = "ADD_PRODUCT_TIER_PRICING_SUCCESS",
  ADD_PRODUCT_TIER_PRICING_FAILED = "ADD_PRODUCT_TIER_PRICING_FAILED",

  UPDATE_PRODUCT_TIER_PRICING_PENDING = "UPDATE_PRODUCT_TIER_PRICING_PENDING",
  UPDATE_PRODUCT_TIER_PRICING_SUCCESS = "UPDATE_PRODUCT_TIER_PRICING_SUCCESS",
  UPDATE_PRODUCT_TIER_PRICING_FAILED = "UPDATE_PRODUCT_TIER_PRICING_FAILED",

  DELETE_VENDOR_PRODUCT_PRICING_PENDING = "DELETE_VENDOR_PRODUCT_PRICING_PENDING",
  DELETE_VENDOR_PRODUCT_PRICING_SUCCESS = "DELETE_VENDOR_PRODUCT_PRICING_SUCCESS",
  DELETE_VENDOR_PRODUCT_PRICING_FAILED = "DELETE_VENDOR_PRODUCT_PRICING_FAILED",

  GET_SINGLE_VENDOR_SUCCESS = "GET_SINGLE_VENDOR_SUCCESS",
  GET_SINGLE_VENDOR_FAILED = "GET_SINGLE_VENDOR_FAILED",
  GET_SINGLE_VENDOR_PENDING = "GET_SINGLE_VENDOR_PENDING",

  ADD_VENDOR_SUCCESS = "ADD_VENDOR_SUCCESS",
  ADD_VENDOR_PENDING = "ADD_VENDOR_PENDING",
  ADD_VENDOR_FAILED = "ADD_VENDOR_FAILED",

  UPDATE_VENDOR_SUCCESS = "UPDATE_VENDOR_SUCCESS",
  UPDATE_VENDOR_PENDING = "UPDATE_VENDOR_PENDING",
  UPDATE_VENDOR_FAILED = "UPDATE_VENDOR_FAILED",

  DELETE_VENDOR_SUCCESS = "DELETE_VENDOR_SUCCESS",
  DELETE_VENDOR_PENDING = "DELETE_VENDOR_PENDING",
  DELETE_VENDOR_FAILED = "DELETE_VENDOR_FAILED",

  UPDATE_VENDOR_STATUS_SUCCESS = "UPDATE_VENDOR_STATUS_SUCCESS",
  UPDATE_VENDOR_STATUS_FAILED = "UPDATE_VENDOR_STATUS_FAILED",
  UPDATE_VENDOR_STATUS_PENDING = "UPDATE_VENDOR_STATUS_PENDING",

  GET_ALL_VENDOR_SUCCESS = "GET_ALL_VENDOR",
  GET_ALL_VENDOR_PENDING = "GET_ALL_VENDOR_PENDING",
  GET_ALL_VENDOR_FAILED = "GET_ALL_VENDOR_FAILED",
}
export default VendorActionTypeEnum;